move feedback link outside of debug panel
