+++ static/css/nl_interface.scss
@@ -132,8 +132,11 @@ $search-section-bottom-height: 7rem;
   }
 }
 
-.nl-query-error,
 .nl-query {
+ padding-top: 2rem; 
+}
+
+.nl-query-error {
   padding: $query-vertical-padding;
 }
 
@@ -157,6 +160,21 @@ $search-section-bottom-height: 7rem;
     position: relative;
     background: $answer-background-color;
     border-radius: $answer-border-radius;
+  }
+
+  .container.feedback-link {
+    background-color: transparent;
+    display: flex;
+    justify-content: flex-end;
+    min-height: 1rem;
+
+    a {
+      color: #70757a;
+      font-style: italic;
+      text-align: right;
+      text-decoration: none;
+      font-size: 0.9rem;
+    }
   }
 
   .block-column {
